Agents-Webservices-FastAPI/requirements.txt
fastapi~=0.116.1
uvicorn[standard]~=0.29.0
langchain-groq~=0.3.6
python-dotenv~=1.1.1
langchain-openai~=0.3.6
pydantic~=2.11.7
pyyaml~=6.0.2
langchain_community~=0.3.24
langchain_huggingface~=0.2.0
chromadb~=1.0.12
chroma-hnswlib~=0.7.6
openai~=1.93.0
langchain-core~=0.3.68
langchain-text-splitters~=0.3.8